services/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
httpx==0.25.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
python-multipart==0.0.6
prometheus-client==0.17.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
redis==5.0.1
starlette-exporter==0.16.0 